move user notifications into sub folder
