+++ packages/openneuro-app/src/scripts/users/notifications/user-notifications-context.tsx
@@ -1,8 +1,8 @@
 import React, { createContext, useCallback, useContext, useState } from "react"
 import * as Sentry from "@sentry/react"
 import { useMutation } from "@apollo/client"
-import { UPDATE_NOTIFICATION_STATUS_MUTATION } from "../queries/datasetEvents"
-import type { MappedNotification } from "../types/event-types"
+import { UPDATE_NOTIFICATION_STATUS_MUTATION } from "../../queries/datasetEvents"
+import type { MappedNotification } from "../../types/event-types"
 
 interface NotificationsContextValue {
   notifications: MappedNotification[]
